Refactor top-level client application code.
Change ClientApplication to extend LibGDX's Game class and then delegate
to EventBasedApplication (the actual, non-LibGDX application), instead
of the other way around. It results in cleaner, smaller, and easier to
understand code.

Rename classes to avoid confusion:

  - Rename ClientApplication to LibGdxGameWrapper because it is not the
    actual application any more, it delegates to the actual application
    (aka wrapping it), giving LibGDX top-level control.

  - Rename ClientApplicationFactory to LibGdxGameFactory because it's
    creating the ApplicationListener / LibGdxGameWrapper that LibGDX
    sees as the actual "game" application that is required to be passed to
    all of the executable sub-projects (android, desktop, html, & ios).

Check Application.shouldShutDown every frame in the
LibGdxGameWrapper.render method, so that the actual application has a
way to flag the wrapper class to shut down the entire LibGDX
application.