Create first version of Play Screen.
The first version of the play screen only shows the background image.

Also add the first version of screen-changing support. ENTER will go to
the next screen, ESCAPE will go to the previous screen. If ENTER is
pressed on the last screen in the sequence, nothing will happen. If
ESCAPE is pressed on the first screen in the sequence, the application
will quit.