First tested version of state machine.
The state machine is not yet wired up to the client / server
architecture yet. It currently works independently via tests:

  ./gradlew core:test

It uses the state machine generator at http://stateforge.com, which
supports hierarchical, parallel, & asynchronous state machines.

The state machine itself is specified via xml, in
core/src/res/GameStateMachine.fsmjava.

GameStateMachine.java & GameModel.java are the java front end of the
state machine system. They interact with GameStateMachine.fsmjava and
with the generated java files in core/src-gen.

GameModel.java specifies the actual actions that will take place in each
state. Theoretically, this class is able to function independently of
the state machine system, without any knowledge of its existence.

GameStateMachine.java receives events from the outside world, and feeds
them into the state machine system by delegating to the generated java
file, GameStateMachineContext.java.

The initial implementation is merely a prototype, or proof-of-concept,
to see how well the StateForge state machine library works.

Only a small fraction of game-related functionality was actually
simulated / mocked-up for the prototype, namely, adding players, and
determining the players' turn order.

The game state machine is currently running asynchronously out of
necessity, since certain events cause other events to be fed back into
the state machine, and many of these need to be scheduled to be dealt
with later rather than immediately, when the state machine is in a valid
state (i.e., has finished dealing with the original event.)

Attempting to run the above scenario synchronously resulted in a
hard-to-debug NullPointerException in the StateForge library, where an
event was attempting to be fed into the state machine when it was in the
middle of a transition. The only solutions were to either set the state
machine to be asynchronous (already supported), or manually schedule
events to run later (which is basically just moving the asynchronous
code implementation burden onto the user, which is stupid when it's
already built in to the library).

The next step is to wire up the prototype to the EventBus, which will
automatically wire it into the client / server system (without having to
know about it).