Make GameModel delegate to PlayerModel.
The GameModel carries out the actions of the state machine. However, it
is too complex to handle all the actions directly, so it needs to
delegate to other sub-models. One such sub-model that has already been
created is PlayerModel.

PlayerModel deals with adding and removing players from the game. It
enforces all the game rules for players, ensuring that only a certain
number of players can join the game, that nothing about them conflicts
with a pre-existing player, that they get assigned an available id,
color, and turn order, etc.

GameModel had been handling this functionality directly in an ad-hoc
way, just to get an initial working version of the state machine.
PlayerModel is way more formal and complete in its handling of players.
This greatly simplifies the code in GameModel.

It is up for debate whether or not PlayerModel should publish events to
the event bus, or leave it up to GameModel to obtain the results of
PlayerModel, and publish events on its behalf depending on those
results, i.e., interpreting the PlayerModel results.

Remove the maxPlayers constructor parameter from PlayerModel, and just
use GameSettings.MAX_PLAYERS everywhere to avoid confusion. It's
designed to be a global constant. It doesn't make sense to have to pass
this global constant in as a parameter rather than just invoking it
directly.

Add a count() method to PlayerModel to get the total player count. Not
sure why this was missing.

Require GameModel to take a PlayerModel parameter in its constructor.
Pass it a PlayerModel instance in GameStateMachineTest.

Add a countdown latch time constant to GameStateMachineTest because if
enough players don't join the game for it to progress, the state machine
will be forever waiting for more players to join. Currently it is set to
5 seconds.

Simulate 50 players attempting to join the game as this is a more robust
test. Some players will be rejected for duplicate names, others will be
rejected because the game is full. Either way, once the game fills up,
it will progress to the DeterminePlayerTurnOrder state, which ends
immediately because it is a placeholder (not yet implemented).

Add javadoc to GameStateMachine.