Add full version of statebuilder-java.
The trial version expired, and in order to use the full version, it has
to be unlocked with a license key. After unlocking, only that jar can be
used, so change it to a local file dependency and put it in core/libs/.
Downloading statebuilder-java from the stateforge maven repository will
no longer work. This way, the full version of the jar will always be
used, and no one has to worry about license keys any more.

The stateforge maven repository can still be used with the statemachine
jar, which apparently does not need to be unlocked, but is required to
use the stateforge library.

Fix the stateforge maven repository url by appending a / to the end.