Minor refactoring of screen functionality.
- Music is started per screen in the show() method, so it should be
stopped in the hide() method. show() is a callback for when the screen
becomes the active screen, and hide() is a callback for when the screen
is no longer the active screen. This way there is some symmetry with the
music starting and stopping.

Rename 'setScreenTo' and 'previousScreenOrDefaultTo' to 'toScreen' and
'toPreviousScreenOr', respectively, which agree better grammatically
with each other and seem easier to read.