Move android manifest info to build.gradle.
It's preferable to give control to the build.gradle instead of having to
use the android manifest xml file. It's more consistent, since the rest
of the build is gradle-based. Also, it makes it easier in the future to
modify these values, such as the version name, either manually, or
programmatically. The chance of forgetting to change the version name is
also higher if it's in the android manifest, since every other version
number change is going to be in a .gradle file.

- Move version code, version name, min sdk version & target sdk version
  from android manifest to android build.gradle.