Add ios/robovm.xml to gitignore.
The xml file should will still be committed to source control. It is
being ignored because every build causes whitespace changes. It is safe
to ignore because it doesn't need to change, and we aren't doing ios dev
at the moment (which is the only thing which might ever cause it to need
to change).