Remove non-color sprites from spriteColorOrder.
spriteColorOrder.txt is confusing because it lists sprites that have no
color analogue; i.e., the HOVERED & SELECTED sprites. Although these
sprites currently exist as part of every country spritesheet, they
cannot be converted to a valid PlayerColor when loaded, causing
problems. For now, simply remove them from the file. A better way will
have to be found to deal with them in the future.