Improve server command line arguments.
Make --countries a required command line argument. Eventually it will
not be passed on the command line at all; it will be calculated at
runtime based on a list of countries in a text file. It is, however,
required in order to calculate the rules of the game, and therefore
currently must be passed in manually.

Print a more user-friendly error message to standard out if the command
line arguments fail, instead of just having the entire application crash
and logging an exception stacktrace. The error message will explain
exactly what went wrong with the command line arguments.