Add PlayMapModel and implement random country assignment.
This commit includes both an initial implementation for PlayMapModel, the model class responsible for mapping countries to player owners, as well as the state implementation for randomly assigning player countries.

It should be noted that PlayMapModel does not care (as I do not believe it should) about whether or not country owners are actually Player objects. It is initialized with an ImmutableSet of Countries; all it cares about is assigning unique Id values to some or all of these Countries. It is entirely agnostic to who or what those Ids represent.

Additional notes:
-Remove StatusEvent (along with its package); all state events now extend from StateEvent and exist in core.model.state.events.
-Add events for player country assignment state; note that all substates (or modes) for country assignment should result in a CountrySelectionCompleteEvent)*
-Add tests for PlayMapModel (test src: core.model.map.PlayMapModelTest).

*CountrySelectionCompleteEvent implementation is incomplete. There needs to be some kind of data attached to it, but I'm not sure what yet.