Fix #268 - robovm.xml bug.
Apparently file trees in groovy are created in different ordering
according to platform (bug?).

This causes spurious changes in robovm.xml when running the gradle build
task on the ios submodule, where only the order of the libs are
changing, and nothing else. This is a problem because robovm.xml must be
checked into source control.

Sort the file tree first to guarantee platform-agnostic ordering.

Update robovm.xml with the correctly-sorted lib order (for the last
time).

Remove robovm.xml from the project's .gitiginore. (It was only there as
a workaround to this bug.)