Fix tank movement bugs.
- Remove continuous tank movement so that the up & down arrows only move
the tank once. Continuous movement was causing a lot of bugs where the
tank could end up going faster than it's normal speed, or not being able
to move at all in a certain direction, or the turret and body would
separate from each other. All of these are now fixed.