Fix settings.txt bug.
GraphicsSettings.INITIAL_WINDOW_WIDTH &
GraphicsSettings.INITIAL_WINDOW_HEIGHT are supposed to be int's. They
get written to ~/peril/settings.txt when it doesn't exist. Because they
were floats the client would crash because it was trying to parse the
window width & height as integers, but couldn't because they were
float's.