Refactor core.net.events packages and Kryonet registration.
Core network events have been reorganized alongside addition of the following new event supertypes: InputRequestEvent and InputResponseEvent.

All concrete implementations of net events are now divided between two parent packages: net.events.client and net.events.server. Events in the client package are sent from the client to the server, and events in the server package are sent from the server to the client.

All interfaces and defualt event implementations that are shared between client and server exist in the 'interfaces' and 'defaults' packages of the parent package (net.events).

InputRequestEvents are sent by the server when input is needed before the game state can progress. For example, the first InputRequestEvent added (but not yet implemented) is PlayerSelectContryInputRequestEvent. This even would be fired upon entry to a waiting state upon which the server must wait for the client to reply with a valid country name that has been selected by the player.

InputResponseEvent is a supertype for a request/success/denied structure (InputResponseRequestEvent, InputResponseSuccessEvent, InputResponseDeniedEvent). InputResponseRequestEvent is an answer from the client to an InputRequest. The input must, however, be validated by the server. As with other types of client requests, the SuccessEvent will be fired upon success, and the DeniedEvent type will be fired upon failure.

Kryonet registration now uses an automated appraoch to registering network classes via Guava reflection. This means all concrete classes in any subpackage of core.net will be automatically located and added to the registration set at runtime. Registered classes outside of the net package still must be added to a constant declared set in the KryonetRegistration class.

Any additional changes are from formatting and/or source clean up.