Fix style issues and refactor country assignment.
PlayMapModel now uses the same 'requestTo*' system as PlayerModel, i.e. instead of assignCountryOwner, it is now requestToAssignCountryOwner which returns a Result. All GameModel now has to do is check the status of the result and use 'failureReasonFrom' to grab the Reason for a Denied event.

Add test cases to PlayMapModelTest to test some of the methods added in this PR (#24).

Add DefaultCountry and CountryFactory to simplify PlayMapModel testing. The number of mocked methods needed to run tests was starting to get out of control versus just writing the implementation.