Fix bug in KryonetServer, minor refactoring.
There is a subtle bug in KryonetServer#start (final int tcpPort) method,
whereby if the TCP port that the server is attempting to bind to throws
an IOException, the Kryonet server thread will not be closed.

This bug occurs because the isRunning flag only gets set to true if the
bind method does not throw; if it does throw, then the stop method
called in the exception handling block returns early because isRunning
is false.

One solution, which was rejected, would be to set the isRunning flag to
true before binding to the port, but this is kind of like lying. The
server isn't really running if it isn't bound to a port. Not to mention
that far in the future some other developer may come along and decide to
move it after the bind method because it doesn't make sense.

The chosen solution is to continue setting the isRunning flag to true
after the bind method, but in the catch block, call super.stop ()
directly, instead of the overridden stop () method. This will bypass the
isRunning flag check and stop the server thread every time. Then, just
to be sure, and to make the code's intention clear, set isRunning to
false in the same catch block, after calling super.stop (), which
parallels the same functionality as the overridden stop () method,
except for the early return from the isRunning flag.

Minor refactoring of KryonetServer & KryonetClient:

- Rename 'listeners' Map to 'networkToKryonetListeners', to adhere to
  project code style.

- Rename add (final NetworkListener listener) method parameter
  'listener' to 'networkListener' to add clarity.

- Rename remove (final NetworkListener listener) method parameter
  'listener' to 'networkListener' to add clarity.

- The last two methods, add/remove are overridden from the Client/Server
  interfaces in fg-tools::net. See
  https://github.com/forerunnergames/fg-tools/commit/435b5c76d357a41e6a451cc6c7229913cd8a1363
  for parallel changes to the actual interface method parameter names.