Implement Continent classes and reinforcement rules.
This commit puts in place some necessary territory and rules framework for moving forward with core game logic.

core-model:
-Add Continent, DefaultContinent, and ContinentFactory.
-Add reinforcement count calculation method to GameRules/ClassicGameRules.
-Add a bit of framework code for game round states.

core-net:
-Add ContinentPacket, and DefaultContinentPacket.

Notes on Continent classes:

Continent is in charge of holding its own configuration data for the territories (country ids) that comprise it as well as the reinforcement bonus for owning it. Letting Continent handle the reinforcement bonus makes more sense both in accordance with the rules of Risk (each continent maps to a specific value) as well as in avoidance of unnecessary complexity in GameRules. "Computing" the value in GameRules would require it to store information about which continents map to what values, which seems out of place for the class. It's cleaner to just configure each Continent as it's created.