Improve PlayMapModel ctor precondition messages.
In core::PlayMapModel, the invalid country count error messages in the
constructor are not very helpful as they don't explain what the actual
minimum and maximum country counts are, nor do they say what country
count was actually passed in.

This made it more difficult than usual to track down a bug while
launching a test server on the command line.

Improve the error messages to output this information.

Remove extra blank line between fields in the same class. Why doesn't
the auto formatter deal with this?