Add CardModel interface and core.model.card components.
CardModel provides an interface for assigning cards from the "deck" to arbitrary (player) Ids. Cards may be assigned until the "hand" of the player is full, or the deck is depleted. The Card deck and player mapping will be handled by package-private types CardDeckHandler and PlayerCardHandler respectively. Cards are traded in via CardSets; a special kind of collection specifically for grouping cards together (like for trade-ins) and checking if the group is a valid match.

The definition of a match (i.e. number of cards required, what combinatorially classifies as a match, etc) is defined by GameRules.

-Add Card, DefaultCard, and CardFactory.
-Add CardModel, PlayerCardHandler, and CardDeckHandler interfaces.
-Add CardSet and CardSet.Match
-Add some Card API rule constraints to GameRules interface.
-Implement changes to GameRules in ClassicGameRules
  --This includes card matching logic (via regex).

Add primary logic tests in PlayerHandlerTest and CardDeckHandler.

Add tests to CardModelTest.

PERIL-115 #done