Implement CardModel API components.
DefaultCardModel is created externally with a GameRules instance (used for state checking) and a set of Cards (the deck).

Internally, DefaultCardModel delegates most of its tasks to DefaultPlayerCardHandler and DefaultCardDealer, both of which are, like their defining interfaces, package-private.
Note: This means that tests and/or alternative package-level implementations of CardModel can use their own versions of those components. In the case of tests, mock or default
objects can be injected into DefaultCardModel.

DefaultCardModel does provide internal logic for 1) checking state to make sure players are allowed to 'add' cards or 'trade-in' cards at any given time and 2) performing trade-ins.

Trade-in matches are precomputed by obtaining the power set of the player's hand and checking each valid subset to see if it qualifies as a "match."

Additional changes:

Fix issue with core build.gradle.

-Fix DefaultCardModel causing tests to fail.
-Add conditional failure assertion to CardModelTest so that FailureReason can be printed on failure.
-Fix core.generateJava not properly deleting old src-gen files; this had been causing me issues when switching to be branches for quite some time now.

Fix CardType.random incorrect upper bound causing sporadic test failures.