Synchronize player box's sorted set for stability.
Use synchronization on the sorted set of players maintained by PlayerBox
in client.

The player box has to deal with players joining and leaving the game in
an asynchronous fashion. For example, while the turn ordered players
field is being iterated (in order to update the display of the player
box), a player could leave the game, causing undefined / unpredictable
results.

Addresses, but does not fix, PERIL-174, which is actually caused by
PERIL-189. This improvement addresses another potential source of bugs
with the player box.