Remove debugging army increment in client.
Remove Debugging Army Increment When Selecting Initial Countries

The client makes up for a lack of core game logic by trying to be smart
and increment the armies on a country after receiving a
PlayerSelectCountryResponseSuccessEvent.

This is really just debugging code, so remove it to get the game
production-ready.

Eventually, the client will respond to an event that changes a country's
armies.

PERIL-180 #Done