Implement Create Game / Host & Play Functionality.
In the client, wire up the Create Multiplayer Game Setup Screen, so that
when clicking the "Create Game" button, it launches a real local host &
play server and then connects to it.

Warning: The server executable must be in the same directory as the
client executable for this to work.

PERIL-154 #Done