Externalize core data.
Move core data (*.txt files) to assets directory, which gets copied to
~/peril/assets.

Country data files need to be in ~/peril/assets. All other map assets are
already in there, and this supports the eventual goal of users being
able to download, create, modify, & load maps into the game. This will
also help support the goal of the client being able to load the country
data to know how many countries are active in a map, in order to be able
to provide valid win percentages on the Create Multiplayer Game: Classic
Mode screen.

Because country data files are being moved here, it doesn't make sense
to some other data files (for continents, cards, etc) in core. These
files will need to be re-created on a per-map basis in the future
anyway.

- Create StreamParserFactory interface to pass to all DataLoader
  implementation constructors, instead of performing a hidden static
  factory call. This allows for injecting different types of
  StreamParserFactory's in different situations.

- Create ExternalStreamParserFactory to create a StreamParser that can
  load data resources external to the jar (i.e. in ~/peril/assets).

- Create InternalStreamParserFactory to create a StreamParser that can
  load data resources internal to the jar (right now this means only
  test resources used by *DataLoaderTest's).

- Remove ModelStreamParserFactory, as it was analogous to
  InternalStreamParserFactory (redundant with a confusing name).

- LibGdxStreamParserFactory also implements StreamParserFactory for
  consistency (it loads assets that are external to the jar, as there
  aren't any production assets inside the jar).

- Add external asset directories in core::AssetSettings for locating
  *.txt data files.

PERIL-217 #Done