Fix GameServerCreator duplicate interface.
GameServerCreator is duplicated in client and core, but the one in core
is unused. GameServerCreator should only exist in core::shared, as it's
not really client's responsibility.

Delete the duplicate GameServerCreator from client, and refactor all
uses to use the one in core.

Move LocalGameServerCreator implementation of GameServerCreator to
core::shared as well, as it's not really client's responsibility.