Dynamically load maps in client.
The Create Multiplayer Game: Classic Mode Setup Screen in the client now
dynamically loads all detected maps in
~/peril/assets/screens/game/play/modes/classic/maps local map assets
directory, and clicking the customize map gear icon (that will
eventually load the Customize Map Popup), cycles automagically to the
next map, re-calculating the valid win percentages because they are
based partially on the map's country count.

Add "my custom map" as a trimmed-down version of the classic map, just
to be able to simulate & test multiple map availability.

Multiple maps are not supported in the core game logic yet though, so
even if a map other than "classic" is chosen on that screen, it will not
be used ("classic" will still be used).

PERIL-210 #Done