Reject clients joining from same ip as server.
We don't want people to easily be able to join an online multiplayer
game with multiple clients on the same machine because it can easily
allow one person controlling multiple clients to easily cheat through
cooperative gameplay. It also makes it appear as if the other clients
are on separate machines.

It violates the idea of a dedicated server, where all the clients are on
separate machines from the server.

server::MultiplayerController checks for matching client addresses in
the JoinGameServerRequestEvent handler, and rejects them with
JoinGameServerDeniedEvent in the case of a match.

Also, server::MultiplayerController will first check for an invalid /
malformed IP address from a client, and reject them with a
JoinGameServerDeniedEvent.

Add tests to server::MultiplayerControllerTest to validate these two new
checks.

PERIL-253 PERIL-251 #Done