Finish multiple map support.
Load map assets for the chosen map on demand instead of at load time:

  - Load minimal map metadata on Create Multiplayer Game: Classic Mode Setup
    Screen, just enough to cycle through a list of map names and update the
    valid win percentages based on the map's country count.

  - Load the remaining assets for the chosen map on the Play Screen after
    creating or joining a game.

  - Assets for non-chosen maps are not loaded.

  - Unload assets for the chosen map after leaving the Play Screen.

Distinguish between stock and custom maps.

Create separate "stock" & "custom" map directories.

Stock map assets are copied to ~/peril/assets in a stock maps directory, but
only the graphical/client assets are treated as authoritative. The *.txt data
files have a redundant, authoritative copy in
core::resources/maps/mode/classic/stock. Only stock maps having a matching
folder name and accompanying data files in core::resources are recognized /
loaded from ~/peril/assets. The map's folder name in core::resources is also a
java package name, so map names containing spaces have underscores there, but
in ~/peril/assets, the underscores become spaces for stock maps.

Custom maps can be created in ~/peril/assets in a separate custom maps
directory. All assets, including *.txt data files, are treated as
authoritative, as core (obviously) cannot account in advance for custom maps,
nor ship them with the game, so the custom map data files on the client must be
authoritative. They are also only "authoritative" when creating a game, not
joining someone else's game. Whoever creates the game, hosts / launches the
server, and that machine's custom map data files are passed to server / core.
This applies for dedicated or host-and-play games.

If a player JOINING a game modifies custom map data files on their local
machine, the client will show the changes, but server / core will ignore the
changes. For example, if a client deletes countries from countries.txt, the
client will not allow that player to select said countries, yet the server will
still see them as valid countries to be selected. This keeps the code simple
because there is no need for trying to enforce / control the integrity of a
non-hosting / non-authoritative client's map data, as it will just be ignored,
so a player will just be causing himself pain & confusion by attempting to
cheat or hack in this way, which is a desirable outcome.

Make the maps directory more accessible in ~/peril/assets by moving it to
~/peril/assets/maps. The directory structure for maps is now:

  ~/peril/assets/maps/classic mode/stock/
      classic/
      classic islands/

  ~/peril/assets/maps/classic mode/custom/
      my uber custom map/

The map name can be passed as --map-name on the command line when running a
dedicated server.

Spaces are allowed in map names. The only place spaces are not allowed is in
the map name as a package name in core::resources/maps/ where underscores are
used in place of spaces.

Use an AssetManager in the client to load & unload graphical assets, which
greatly facilitates dynamic loading of map assets, and makes having loading
screens in the future possible.

New maps:

To show off the multiple map support and prove that it works, there are
now Classic & Classic Islands stock maps, that are official. There is
also a fake custom map that really works, called My Uber Custom Map. :P

PERIL-245 PERIL-246 PERIL-247 PERIL-263 #Done