Add game setup settings to settings.txt.
Add the following settings to settings.txt:

player-name
clan-tag
server-address
server-title
players-classic-mode
spectators
map-name-classic-mode
initial-country-assignment-classic-mode
win-percent-classic-mode

The values are parsed early as much as possible before launching the
client.  If a setting is still invalid, such as win-percent-classic-mode
(because it cannot be validated until the map is loaded), or
map-name-classic-mode, an error popup will appear on the setup screen
explaining what went wrong and to check the settings file.

The new values from settings.txt are injected into the setup screens.
This greatly speeds up client debugging time, as the values do not have
to be entered manually every single time.

Miscellaneous:

- Map names can now have numbers, and must have a minimum of two
  characters.

PERIL-177 #Done