Implement reinforcement phase core logic.
Upon the beginning of a player's turn, a BeginReinforcementPhaseEvent will be sent to all players containing a PlayerPacket representing the current player reinforcing.

Country reinforcements are then calculated and placed into the players hand. All players are notified via a PlayerArmiesChangedEvent.

The current player is sent a PlayerReinforceCountriesRequestEvent, which contains a pre-computed combinatorial set of card matches for trade-in, the player's owned countries, and a boolean condition for whether or not trade-in is immediately required.

PlayerReinforceCountriesRequestEvent also contains the reinforcement bonus values given by countries, continents, and the next trade-in bonus.

PlayerReinforceCountries[Response]RequestEvent(s) will only accept a single trade-in, since only one will be possible during the reinforcement phase.

The player's response is checked to verify the trade-in (if any) and the validity of the requested country reinforcement counts. A PlayerReinforceCountriesResponseSuccess/DeniedEvent is sent after validation is complete.

BeginReinforcementPhaseEvent now only contains a PlayerPacket.

Update GameModel/GameModelTest to work with changes.

Additional changes:
 - Add additional utility method to PlayMapModel/DefaultPlayMapModel.
 - Add toString to DefaultCardSetPacket.
 - Fix CardDealerTest.generateTestCards() using bad value cycle for card types; caused far too many WILDCARD types to be generated and no TYPE3.
 - Add tests to GameModelTest for new reinforcement phase methods.
 - Add getContinentsOwnedBy method to PlayMapModel/DefaultPlayMapModel.

PERIL-110 #Done