Remove unreliable test.
MultiplayerControllerTest#testClientJoinRequestDeniedBecauseInvalidIpLetters
fails sometimes depending on local nameserver configuration. A new
InetAddress is created with address string "invalidip1234", which
performs a dns lookup to attempt to resolve it as a hostname. This
violates the intention of the test, which is intended to test for
rejection of clients having invalid ip addresses in
MultiplayerController#handleEvent (JoinGameServerRequestEvent event, Remote event),
more specifically, it is intended to trigger the following block of code
in said method:

    if (!NetworkConstants.isValidIpAddress (client.getAddress ()))
    {
      sendJoinGameServerDenied (client, "Your IP address [" + client.getAddress () + "] is invalid.");
      return;
    }

Unfortunately, by the time this code is reached in the test, in some
cases the client has a valid placeholder ip address returned by the
local nameserver, i.e., the ip address of the nameserver itself.

Remove the test for now.

PERIL-288 #Done