Improve event bus error handling.
EventBusFactory can now take a variable number of DeadEventHandler's,
that respond to DeadMessage's (Event's that no @Handler methods are
listening for).

Add EventBusFactory#create method overloads for taking
DeadEventHandler's.

LoggingDeadEventHandler simply logs a warning about the DeadMessage.

Create AssetLoadingErrorDeadEventHandler on the client for handling
DeadMessage's containing an AssetLoadingErrorEvent, which throws a
RuntimeException for any AssetLoadingErrorEvent (which contains a
Throwable) that is not handled elsewhere.  This is important, for
example, to cause the client to fully crash when an asset that is not
yet needed fails to load, so that the main window will be destroyed, and
a crash file generated.

Create ThrowingErrorPublicationHandler in common::eventbus for re-throwing
Throwable's caught by the event bus, wrapped in a RuntimeException.

Pass instances of LoggingDeadEventHandler &
AssetLoadingErrorDeadEventHandler to EventBusFactory in
LibGdxGameFactory on the client.

Pass instance of ThrowingErrorPublicationHandler to EventBusFactory in
LibGdxGameFactory on the client, so that when
AssetLoadingErrorDeadEventHandler throws an exception, it will not be
swallowed, and will cause the client application to crash, generating a
crash file and destroying the main window.