Rename DEFAULT_S3_BUCKET_NAME...
... to DEFAULT_S3_BUCKET_PATH.

S3_BUCKET_NAME used to be just "assets.peril.forerunnergames.com", but
now an s3:// path prefix is required, making it an actual path, not just
a name. It is already referred to everywhere else as a path, and in some
cases the path prefix is stripped before use, changing it to a name.
Rename to S3_BUCKET_PATH to avoid any confusion.