Fix client asset loading bug.
Client crashes with stack trace:

java.lang.IllegalStateException: Cannot get asset [cursors/normalCursor.png] because it was not loaded first by MultiSourceAssetManager.
	at com.forerunnergames.peril.client.assets.MultiSourceAssetManager.get(MultiSourceAssetManager.java:57) ~[peril-client-0.1-SNAPSHOT.jar:na]
	at com.forerunnergames.peril.client.assets.MultiSourceAssetManager.get(MultiSourceAssetManager.java:44) ~[peril-client-0.1-SNAPSHOT.jar:na]
	at com.forerunnergames.peril.client.ui.widgets.AbstractWidgetFactory.getAsset(AbstractWidgetFactory.java:475) ~[peril-client-0.1-SNAPSHOT.jar:na]
	at com.forerunnergames.peril.client.ui.widgets.AbstractWidgetFactory.createNormalCursor(AbstractWidgetFactory.java:329) ~[peril-client-0.1-SNAPSHOT.jar:na]
	at com.forerunnergames.peril.client.ui.screens.loading.MenuToPlayLoadingScreen.<init>(MenuToPlayLoadingScreen.java:131) ~[peril-client-0.1-SNAPSHOT.jar:na]
	at com.forerunnergames.peril.client.ui.screens.ScreenFactory.create(ScreenFactory.java:123) ~[peril-client-0.1-SNAPSHOT.jar:na]
	at com.forerunnergames.peril.client.ui.screens.ScreenController.toScreen(ScreenController.java:108) ~[peril-client-0.1-SNAPSHOT.jar:na]
	at com.forerunnergames.peril.client.ui.screens.menus.AbstractMenuScreen.toScreen(AbstractMenuScreen.java:395) ~[peril-client-0.1-SNAPSHOT.jar:na]
	at com.forerunnergames.peril.client.ui.screens.menus.multiplayer.modes.classic.creategame.MultiplayerClassicGameModeCreateGameMenuScreen.access$1700(MultiplayerClassicGameModeCreateGameMenuScreen.java:59) ~[peril-client-0.1-SNAPSHOT.jar:na]
	at com.forerunnergames.peril.client.ui.screens.menus.multiplayer.modes.classic.creategame.MultiplayerClassicGameModeCreateGameMenuScreen$5.clicked(MultiplayerClassicGameModeCreateGameMenuScreen.java:341) ~[peril-client-0.1-SNAPSHOT.jar:na]
	at com.badlogic.gdx.scenes.scene2d.utils.ClickListener.touchUp(ClickListener.java:89) ~[peril-client-0.1-SNAPSHOT.jar:na]
	at com.badlogic.gdx.scenes.scene2d.InputListener.handle(InputListener.java:58) ~[peril-client-0.1-SNAPSHOT.jar:na]
	at com.badlogic.gdx.scenes.scene2d.Stage.touchUp(Stage.java:348) ~[peril-client-0.1-SNAPSHOT.jar:na]
	at com.badlogic.gdx.InputMultiplexer.touchUp(InputMultiplexer.java:96) ~[peril-client-0.1-SNAPSHOT.jar:na]
	at com.badlogic.gdx.backends.lwjgl.LwjglInput.processEvents(LwjglInput.java:316) ~[peril-client-0.1-SNAPSHOT.jar:na]
	at com.badlogic.gdx.backends.lwjgl.LwjglApplication.mainLoop(LwjglApplication.java:207) ~[peril-client-0.1-SNAPSHOT.jar:na]
	at com.badlogic.gdx.backends.lwjgl.LwjglApplication$1.run(LwjglApplication.java:120) ~[peril-client-0.1-SNAPSHOT.jar:na]

cursors/normalCursor.png is being unloaded at the end of the splash
screen because it is listed as a splash screen asset
(AssetSettings.SPLASH_SCREEN_ASSET_DESCRIPTORS). The problem is that
first, it has to be pre-loaded before the splash screen is displayed,
then assets are updated, then it is loaded as an always-loaded assets
(AssetSettings.ALWAYS_LOADED_ASSET_DESCRIPTORS), which is part of the
initially-loaded assets (AssetSettings.INITIAL_ASSET_DESCRIPTORS) loaded
by the splash screen.

This happens because the normal cursor asset is used on the splash
screen, but is also permanently loaded for the entire application's
lifetime.

The solution is to differentiate between the assets that need to be
pre-loaded before the splash screen is displayed, and the assets that
need to be UNLOADED after the splash screen is complete, and ready to
transition to the start screen.

Pre-load the normal cursor asset, but don't unload it at the end of the
splash screen.

Note: The normal cursor asset also gets re-loaded during the splash
screen, after assets are updated, so the normal cursor gets a chance to
have itself updated for the remainder of the application, after the
splash screen ends.