Implement attack phase game state.
*Add battle APIs to core (BattleModel, BattleResult, BattleActor, etc).*

BattleModel handles the validation of attack request data and generation of battle results. After a request has been validated, a pending AttackOrder representing the request is returned. Once the relevant data has been obtained from the defending player, the AttackOrder can be processed by BattleModel to obtain the BattleResult.

The defending player is sent a PlayerDefendCountryRequestEvent after validation of the attacking player's request. The PlayerDefendCountryResponseRequestEvent contains the number of dice the defender is choosing to roll. Once the response is validated, a battle result is generated and sent to all players in the form of a PlayerAttackCountrySuccessEvent.

*Add PlayerTurnDataCache*

PlayerTurnDataCache provides a mechanism for persisting data across state transitions in GameModel. Currently, it is intended specifically for persisting player turn data (i.e. pending attack orders, battle data, etc). Its use could be expanded in the future to tie together other data processing stages of the state machine.

*Add PlayerRequestEvent, PlayerSuccessEvent, etc*

It is necessary for core to be able to validate from which player client request events originate, mostly in order to verify that players aren't submitting valid request events at invalid times (i.e. during another player's turn). These events as well as their handlers in server/core help to provide that functionality.

*Other changes*
- Add CountryOwnerChangedEvent.
- Update CountryArmiesChangedEvent to use CountryPacket instead of country name string.
- Update server to properly handle player success and request events.
- Update a few client classes to reflect changes to CountryArmiesChangedEvent.

PERIL-322 #Done